refactor/point: Enum 패키지 생성 및 Enum 클래스 분리 (create Enum package and relocate Enum class)
